refractor(structure): reorganise files and lint code
- Moved files to more meaningful files
- Refractored code for improved readability and maintainability
- Linted all files to ensure consistent style
