+++ README.md
@@ -17,4 +17,3 @@ settings.
 ## TODO
 
 - [X] An online demo.
-- [ ] Add video references at `src/app/page.tsx` rather than in the component itself
