--- src/components/VideoPlayer/VideoPlayer.module.css
@@ -1,255 +1,0 @@
-.pointer {
-  cursor: pointer;
-  margin: 0 1rem;
-}
-
-.videoContainer {
-  width: 1000px;
-  border-radius: 10px;
-  aspect-ratio: 16/9;
-  overflow: hidden;
-  position: relative;
-  max-width: 100%;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.video {
-  width: 100%;
-  height: 100%;
-  position: relative;
-}
-.videoContainer.fullScreenVideo {
-  width: 100%;
-  height: 100%;
-  border-radius: 0;
-}
-
-.videoControlsContainer.hidden {
-  visibility: hidden;
-  opacity: 0;
-}
-.videoContainer.hidden {
-  cursor: none;
-}
-.videoControlsContainer {
-  transition:
-    opacity 300ms ease,
-    visibility 300ms ease;
-  z-index: 2;
-  position: absolute;
-  display: block;
-  bottom: 0;
-  width: 100%;
-  padding: 0.5rem 1rem 1rem;
-  color: white;
-}
-.videoControlsContainer:hover {
-  visibility: visible !important;
-  opacity: 1 !important;
-}
-.videoControlsContainer::before {
-  z-index: -1;
-  content: "";
-  width: 100%;
-  height: 30vh;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  background: linear-gradient(transparent, #000000a6);
-  pointer-events: none;
-}
-
-.controls {
-  display: flex;
-  height: 3rem;
-  position: relative;
-  align-items: center;
-}
-.controls > * {
-  max-height: 3rem;
-}
-.time {
-  display: inline-block;
-  min-width: 4rem;
-  max-width: 3rem;
-  -webkit-touch-callout: none;
-  -webkit-user-select: none;
-  -khtml-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-}
-.controlBtn {
-  display: inline-block;
-  min-height: 3rem;
-  min-width: 3rem;
-  cursor: pointer;
-  background-color: #0000;
-  border: none;
-  outline: none;
-}
-.controlBtn svg {
-  fill: #e7e7e7;
-  font-size: 0.8em;
-  transform: scale(0.6);
-}
-#playPauseBtn svg {
-  transform: scale(0.8);
-}
-#volume {
-  margin-right: 1rem;
-  cursor: default;
-  position: relative;
-}
-.volCont {
-  position: relative;
-  max-height: 3rem;
-}
-.volCont > div#volSlider {
-  position: absolute;
-  bottom: 40px;
-  left: 10px;
-  width: 30px;
-  height: 0;
-  transition: all 250ms linear;
-  opacity: 0.4;
-  background-color: #00000050;
-  border-radius: 3px;
-  overflow: hidden;
-}
-#volRange {
-  position: absolute;
-  transform: rotate(-90deg);
-  transform-origin: left;
-  bottom: 0px;
-  left: 50%;
-  width: 100px;
-  cursor: pointer;
-}
-.volCont:hover > div#volSlider {
-  height: 115px;
-  border: 1px solid #e3e3e38e;
-  opacity: 1;
-}
-#fullScreen > svg {
-  fill: none;
-  transform: scale(0.6);
-}
-
-.timelineContainer {
-  position: relative;
-  width: 100%;
-  margin: 0 15px;
-  padding: 5px 0;
-}
-.timeline {
-  transition: all 200ms ease-in-out;
-  --progress: 0%;
-  position: relative;
-  width: 100%;
-  height: 3px;
-  background-color: #eeeeee6e;
-  z-index: 2;
-  border-radius: 5px;
-  cursor: pointer;
-  /* box-shadow: 1px 1px 5px #b8b8b8cb; */
-}
-.timeline::before {
-  content: "";
-  z-index: 4;
-  position: absolute;
-  height: 100%;
-  left: 0;
-  width: var(--progress, 0%);
-  background-color: #e22121;
-  border-radius: 5px;
-  transition: height 200ms linear;
-}
-.timeline:hover,
-.timelineContainer:hover .timeline {
-  height: 5px;
-}
-.timeline::after {
-  --size: 25px;
-  transition: scale 200ms ease;
-  content: "";
-  position: absolute;
-  width: var(--size, 0);
-  height: var(--size, 0);
-  cursor: pointer;
-  bottom: calc(50% - var(--size) / 2);
-  left: calc(var(--progress) - (var(--size) / 2));
-  scale: 0;
-  border-radius: 50%;
-  background-color: #e22121;
-}
-.timelineContainer:hover .timeline::after,
-.scrubbing.timeline::after {
-  scale: 1;
-}
-
-/* visual helpers */
-.helperCont {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  pointer-events: none;
-}
-.helper {
-  margin: auto;
-  opacity: 1;
-  transform: scale(1);
-  background-color: #00000050;
-  width: 4.5rem;
-  height: 4.5rem;
-  border-radius: 50%;
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 2;
-  font-size: 2rem;
-}
-/* .helper > i#playHelper {
-}
-.helper > i#volOffHelper {
-  width: 1.7rem;
-} */
-.helper > svg {
-  fill: #e7e7e7;
-}
-.helper.show {
-  opacity: 1;
-}
-.helper.hide {
-  opacity: 0;
-  transform: scale(1.5);
-  transition: all 300ms ease-in-out;
-}
-.helperCont > span#volLabelHelper {
-  --vol-label-size: 5rem;
-  display: inline-block;
-  opacity: 0;
-  position: absolute;
-  bottom: 90%;
-  left: calc(50% - var(--vol-label-size) / 2);
-  width: var(--vol-label-size);
-  background-color: #00000090;
-  text-align: center;
-  font-size: 1.5rem;
-  padding: 0.4rem;
-  border-radius: 5px;
-  z-index: 2;
-}
-.helperCont > span#volLabelHelper.show {
-  opacity: 1;
-}
-.helperCont > span#volLabelHelper.hide {
-  opacity: 0;
-  transition: all 1000ms cubic-bezier(0.95, 0.05, 0.795, 0.035);
-}
