+++ src/components/video/Controls/styles.module.css
@@ -1,0 +1,43 @@
+
+.videoControlsContainer.hidden {
+  visibility: hidden;
+  opacity: 0;
+}
+
+.videoControlsContainer {
+  transition:
+    opacity 300ms ease,
+    visibility 300ms ease;
+  z-index: 2;
+  position: absolute;
+  display: block;
+  bottom: 0;
+  width: 100%;
+  padding: 0.5rem 1rem 1rem;
+  color: white;
+}
+.videoControlsContainer:hover {
+  visibility: visible !important;
+  opacity: 1 !important;
+}
+.videoControlsContainer::before {
+  z-index: -1;
+  content: "";
+  width: 100%;
+  height: 30vh;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  background: linear-gradient(transparent, #000000a6);
+  pointer-events: none;
+}
+
+.controls {
+  display: flex;
+  height: 3rem;
+  position: relative;
+  align-items: center;
+}
+.controls > * {
+  max-height: 3rem;
+}
