+++ src/components/video/VisualHelper/styles.module.css
@@ -1,0 +1,59 @@
+/* visual helpers */
+.helperCont {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  pointer-events: none;
+}
+.helper {
+  margin: auto;
+  opacity: 1;
+  transform: scale(1);
+  background-color: #00000050;
+  width: 4.5rem;
+  height: 4.5rem;
+  border-radius: 50%;
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 2;
+  font-size: 2rem;
+}
+
+.helper > svg {
+  fill: #e7e7e7;
+}
+.helper.show {
+  opacity: 1;
+}
+.helper.hide {
+  opacity: 0;
+  transform: scale(1.5);
+  transition: all 300ms ease-in-out;
+}
+.helperCont > span#volLabelHelper {
+  --vol-label-size: 5rem;
+  display: inline-block;
+  opacity: 0;
+  position: absolute;
+  bottom: 90%;
+  left: calc(50% - var(--vol-label-size) / 2);
+  width: var(--vol-label-size);
+  background-color: #00000090;
+  text-align: center;
+  font-size: 1.5rem;
+  padding: 0.4rem;
+  border-radius: 5px;
+  z-index: 2;
+}
+.helperCont > span#volLabelHelper.show {
+  opacity: 1;
+}
+.helperCont > span#volLabelHelper.hide {
+  opacity: 0;
+  transition: all 1000ms cubic-bezier(0.95, 0.05, 0.795, 0.035);
+}
